requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.2
discord.py==1.7.3
dnspython==1.16.0
idna==3.2
multidict==5.1.0
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.2
pycodestyle==2.7.0
pymongo==3.12.0
requests==2.26.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
